ai_lecture_2/requirements.txt
langchain==0.1.10
langgraph==0.0.30
langchain-openai==0.0.5
python-dotenv==1.0.1
pydantic==2.6.1

